requirements.txt
# Network Automation Dependencies
netmiko>=4.1.0
paramiko>=3.0.0
pyyaml>=6.0
jinja2>=3.1.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Dependencies
black>=23.0.0
flake8>=6.0.0

